+++ app/GraphNode.svelte
@@ -6,18 +6,19 @@
 
     let context = false;
     $: context = $currentContext?.type == "Revision" && header == $currentContext.header;
+    $: wcClass = header.working_copy_of != null ? "other-wc" : "wc";
 </script>
 
 {#if header.is_immutable}
     {#if header.is_working_copy}
-        <circle class="wc" class:context cx="9" cy="15" r="6" />
+        <circle class={wcClass} class:context cx="9" cy="15" r="6" />
     {:else}
         <circle class:context cx="9" cy="15" r="6" />
     {/if}
 {:else}
     <circle class:context cx="9" cy="15" r="6" class="mutable" />
     {#if header.is_working_copy}
-        <circle class="wc" class:context cx="9" cy="15" r="3" />
+        <circle class={wcClass} class:context cx="9" cy="15" r="3" />
     {/if}
 {/if}
 
@@ -31,6 +32,11 @@
     .wc {
         stroke: var(--ctp-green);
         fill: var(--ctp-green);
+    }
+
+    .other-wc {
+        stroke: var(--ctp-yellow);
+        fill: var(--ctp-yellow);
     }
 
     .context {