+++ src/messages/queries.rs
@@ -94,6 +94,7 @@ pub struct RevHeader {
     pub author: RevAuthor,
     pub has_conflict: bool,
     pub is_working_copy: bool,
+    pub working_copy_of: Option<String>,
     pub is_immutable: bool,
     pub refs: Vec<StoreRef>,
     pub parent_ids: Vec<CommitId>,